Data of MK-ESPDMF
The MK-ESPDMF model utilized a single piece of submersible electric pump well failure data in its testing, which originated from an offshore oil extraction plant in China. To prevent data leakage, only one piece of failure data was provided. The related research paper has been submitted to the Journal of Petroleum Exploration and Production Technology. The failure is a broken shaft between the upper pump and the gas-liquid pump.